refactor: separate RestmapConf and WebConf into separate files
